Move general arcgis rest utility functions to core
