+++ src/providers/arcgisrest/qgsarcgisrestproviderutils.h
@@ -1,0 +1,65 @@
+/***************************************************************************
+    qgsarcgisrestproviderutils.h
+    ----------------------------
+    begin                : Nov 25, 2015
+    copyright            : (C) 2015 by Sandro Mani
+    email                : manisandro@gmail.com
+ ***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ ***************************************************************************/
+#ifndef QGSARCGISRESTPROVIDERUTILS_H
+#define QGSARCGISRESTPROVIDERUTILS_H
+
+
+#include <QStringList>
+
+#include "qgsrenderer.h"
+
+class QNetworkReply;
+
+
+class QgsArcGisAsyncQuery : public QObject
+{
+    Q_OBJECT
+  public:
+    QgsArcGisAsyncQuery( QObject *parent = nullptr );
+    ~QgsArcGisAsyncQuery() override;
+
+    void start( const QUrl &url, const QString &authCfg, QByteArray *result, bool allowCache = false, const QgsStringMap &headers = QgsStringMap() );
+  signals:
+    void finished();
+    void failed( QString errorTitle, QString errorName );
+  private slots:
+    void handleReply();
+
+  private:
+    QNetworkReply *mReply = nullptr;
+    QByteArray *mResult = nullptr;
+};
+
+class QgsArcGisAsyncParallelQuery : public QObject
+{
+    Q_OBJECT
+  public:
+    QgsArcGisAsyncParallelQuery( const QString &authcfg, const QgsStringMap &requestHeaders, QObject *parent = nullptr );
+    void start( const QVector<QUrl> &urls, QVector<QByteArray> *results, bool allowCache = false );
+
+  signals:
+    void finished( QStringList errors );
+  private slots:
+    void handleReply();
+
+  private:
+    QVector<QByteArray> *mResults = nullptr;
+    int mPendingRequests = 0;
+    QStringList mErrors;
+    QString mAuthCfg;
+    QgsStringMap mRequestHeaders;
+};
+
+#endif // QGSARCGISRESTUTILS_H
